% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{ActivityParams}
\alias{ActivityParams}
\title{Create ActivityParams object}
\usage{
ActivityParams(lengthCriticalRegion = 10, probInactiveWhenMutated = 0.01)
}
\arguments{
\item{lengthCriticalRegion}{Sequence length (number of nucleotides) of the critical region of a retrotransposon}

\item{probInactiveWhenMutated}{The probability that a singe point mutation to the critical region causes a sequence to become inactive}
}
\description{
Create ActivityParams object
}
