% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{FamilyParams}
\alias{FamilyParams}
\title{Create FamilyParams object}
\usage{
FamilyParams(familyCoherence = 0.7, maxFamilyRepresentatives = 20)
}
\arguments{
\item{familyCoherence}{What sequence similarity do two sequences have to be to each other for them to be considered to be of the same family?}

\item{maxFamilyRepresentatives}{How many family representatives to keep track of during the simulation?}
}
\value{
A bundling of the parameters given to it as a FamilyParams object
}
\description{
Create FamilyParams object
}
\examples{
familyParams <- FamilyParams(familyCoherence = 0.775)
}
