% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.ActivityParams}
\alias{is.ActivityParams}
\title{Is x an ActivityParams object?}
\usage{
is.ActivityParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a ActivityParams object, and FALSE otherwise
}
\description{
Is x an ActivityParams object?
}
\examples{
is.ActivityParams(ActivityParams()) # TRUE
is.ActivityParams(data.frame()) # FALSE
}
