% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.BurstParams}
\alias{is.BurstParams}
\title{Is x a BurstParams object?}
\usage{
is.BurstParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a BurstParams object, and FALSE otherwise
}
\description{
Is x a BurstParams object?
}
\examples{
is.BurstParams(BurstParams()) # TRUE
is.BurstParams(data.frame()) # FALSE
}
