% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.MutationParams}
\alias{is.MutationParams}
\title{Is x a MutationParams object?}
\usage{
is.MutationParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a MutationParams object, and FALSE otherwise
}
\description{
Is x a MutationParams object?
}
\examples{
is.MutationParams(MutationParams()) # TRUE
is.MutationParams(data.frame()) # FALSE
}
