% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.OutputParams}
\alias{is.OutputParams}
\title{Is x a OutputParams object?}
\usage{
is.OutputParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a OutputParams object, and FALSE otherwise
}
\description{
Is x a OutputParams object?
}
\examples{
is.OutputParams(OutputParams()) # TRUE
is.OutputParams(data.frame()) # FALSE
}
