% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.RecombParams}
\alias{is.RecombParams}
\title{Is x a RecombParams object?}
\usage{
is.RecombParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a RecombParams object, and FALSE otherwise
}
\description{
Is x a RecombParams object?
}
\examples{
is.RecombParams(RecombParams()) # TRUE
is.RecombParams(data.frame()) # FALSE
}
