% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.SelectionParams}
\alias{is.SelectionParams}
\title{Is x a SelectionParams object?}
\usage{
is.SelectionParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a SelectionParams object, and FALSE otherwise
}
\description{
Is x a SelectionParams object?
}
\examples{
is.SelectionParams(SelectionParams()) # TRUE
is.SelectionParams(data.frame()) # FALSE
}
