% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.SequenceParams}
\alias{is.SequenceParams}
\title{Is x a SequenceParams object?}
\usage{
is.SequenceParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a SequenceParams object, and FALSE otherwise
}
\description{
Is x a SequenceParams object?
}
\examples{
is.SequenceParams(SequenceParams()) # TRUE
is.SequenceParams(data.frame()) # FALSE
}
