% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.SimulationParams}
\alias{is.SimulationParams}
\title{Is x a SimulationParams object?}
\usage{
is.SimulationParams(x)
}
\arguments{
\item{x}{An arbitrary object}
}
\value{
A logical, TRUE if \code{x} is a SimulationParams object, and FALSE otherwise
}
\description{
Is x a SimulationParams object?
}
\examples{
is.SimulationParams(SimulationParams()) # TRUE
is.SimulationParams(data.frame()) # FALSE
}
