% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_family_distances}
\alias{plot_family_distances}
\title{Plot pairwise distance within families}
\usage{
plot_family_distances(data)
}
\arguments{
\item{data}{Simulation data obtained using `input_data`}
}
\description{
Plot pairwise distance within families
}
