% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_initial_distances}
\alias{plot_initial_distances}
\title{Plot distance to initial sequence}
\usage{
plot_initial_distances(data)
}
\arguments{
\item{data}{Simulation data obtained using `input_data`}
}
\description{
Plot distance to initial sequence
}
