% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Plot mosaic of recombination}
\usage{
plot_phylogeny(data)
}
\arguments{
\item{data}{Simulation data obtained using `input_data`}
}
\description{
Plot mosaic of recombination
}
