% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_phylogeny_mosaic}
\alias{plot_phylogeny_mosaic}
\title{Plot mosaic of recombination}
\usage{
plot_phylogeny_mosaic(data)

plot_phylogeny_mosaic(data)
}
\arguments{
\item{data}{Simulation data obtained using `input_data`}
}
\description{
Plot mosaic of recombination

Plot age of families
}
