% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plotInitialDistance}
\alias{plotInitialDistance}
\title{Plotting function wrapper}
\usage{
plotInitialDistance(data)
}
\arguments{
\item{data}{Simulation output data obtained from parseEvolutionOutput}
}
\description{
Plotting function wrapper
}
