% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plotPairwiseDistance}
\alias{plotPairwiseDistance}
\title{Plotting function wrapper}
\usage{
plotPairwiseDistance(data)
}
\arguments{
\item{data}{Simulation output data obtained from parseEvolutionOutput}
}
\description{
Plotting function wrapper
}
