% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcombinator.R
\docType{package}
\name{rcombinator-package}
\alias{rcombinator}
\alias{rcombinator-package}
\title{rcombinator: Simulating The Evolution of Retrotransposons Over Time}
\description{
This software package performs simulations on retrotransposon sequences over
long periods of time (~500M years). It takes into account bursts and
recombination events. Furthermore, it has functions to perform basic data
analysis on the simulation results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sams25/rcombinator}
  \item Report bugs at \url{https://github.com/sams25/rcombinator/issues}
}

}
\author{
\strong{Maintainer}: Anindya Sharma \email{anindya.sharma1@gmail.com}

}
