% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{RecombParams}
\alias{RecombParams}
\title{Create RecombParams object}
\usage{
RecombParams(recombMean = 1.5, recombSimilarity = 0.8)
}
\arguments{
\item{recombMean}{The expected number of template switches during recombination between two sequences (chosen from a Poisson distribution with this as its mean)}

\item{recombSimilarity}{How similar does an active sequence have to be with another sequence for them to be allowed to be recombine during transposition?}
}
\value{
A bundling of the parameters given to it as a RecombParams object
}
\description{
Create RecombParams object
}
\examples{
recombParams <- RecombParams(recombMean = 2)
}
