% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrocombinator.R
\docType{package}
\name{retrocombinator-package}
\alias{retrocombinator}
\alias{retrocombinator-package}
\title{retrocombinator: Simulating The Evolution of Retrotransposons Over Time}
\description{
This software package performs simulations on retrotransposon sequences over
long periods of time (~500M years). It takes into account bursts and
recombination events. Furthermore, it has functions to perform basic data
analysis on the simulation results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/drostlab/retrocombinator}
  \item Report bugs at \url{https://github.com/drostlab/retrocombinator/issues}
}

}
\author{
\strong{Maintainer}: Anindya Sharma \email{anindya.sharma1@gmail.com}

Authors:
\itemize{
  \item Hajk-Georg Drost \email{hgd23@cam.ac.uk}
}

}
