% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SeedParams}
\alias{SeedParams}
\title{Create SeedParams object}
\usage{
SeedParams(to_randomise = FALSE, to_seed = TRUE, seed = 0)
}
\arguments{
\item{to_randomise}{Should this simulation be run with a random seed to begin
with? (Based on system time)}

\item{to_seed}{Should this sequence be run with a specific initial seed? If
so, then the seed is specified by the parameter sseed}

\item{seed}{Refer to documentation of to_seed}
}
\description{
Create SeedParams object
}
