% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SelectionParams}
\alias{SelectionParams}
\title{Create SelectionParams object}
\usage{
SelectionParams(selectionThreshold = 0.3)
}
\arguments{
\item{selectionThreshold}{What sequence similarity to the original sequence do we wish to maintain? Sequences that diverge beyond this similarity level are dropped over the course of simulation}
}
\description{
Create SelectionParams object
}
