% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SequenceParams}
\alias{SequenceParams}
\title{Create SequenceParams object}
\usage{
SequenceParams(
  numInitialCopies = 100,
  sequenceLength = 500,
  initialSequence = NULL
)
}
\arguments{
\item{numInitialCopies}{The initial number of copies of a given sequence to consider}

\item{sequenceLength}{The number of nucleotides in each of the sequences in the
simulation}

\item{initialSequence}{An initial sequence (ignores sequenceLength if this is
specified)}
}
\description{
Create SequenceParams object
}
