% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run an entire simulation with flags}
\usage{
simulate(
  sequenceParams,
  simulationParams,
  mutationParams,
  flagParams,
  burstParams,
  speciationParams,
  outputParams,
  seedParams
)
}
\description{
Run an entire simulation with flags
}
