% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_with_flags_random}
\alias{simulate_with_flags_random}
\title{Run an entire simulation with flags}
\usage{
simulate_with_flags_random(num_seq, seq_length, point_mutation_model = "K80",
  num_sensitive_posns = 10, inactive_probability = 0.1, num_jumps = 100,
  timestep = 1, burst_probability = 0.1, burst_mean = 1,
  max_active_copies = 100, max_total_copies = 200, recomb_mean = 1,
  selection_threshold = 0, fam_proportion = 1, fam_percentage = 100,
  file_out, num_out_tags = 100, num_out_init = 100, num_out_seqs = 10,
  num_out_pair = 20, to_randomise = TRUE, to_seed = FALSE, seed = 0,
  sequence_numbering = 0, family_numbering = 0)
}
\description{
Run an entire simulation with flags
}
