% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEvolution.R
\name{simulateEvolution}
\alias{simulateEvolution}
\title{Run an entire simulation with flags}
\usage{
simulateEvolution(
  sequenceParams = SequenceParams(),
  activityParams = ActivityParams(),
  mutationParams = MutationParams(),
  burstParams = BurstParams(),
  recombParams = RecombParams(),
  selectionParams = SelectionParams(),
  familyParams = FamilyParams(),
  simulationParams = SimulationParams(),
  outputParams = OutputParams(),
  seedParams = SeedParams()
)
}
\description{
Run an entire simulation with flags
}
