% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SpeciationParams}
\alias{SpeciationParams}
\title{Create SpeciationParams object}
\usage{
SpeciationParams(
  selectionThreshold = 0.5,
  speciesSimilarity = 0.7,
  speciesCoherence = 0.5
)
}
\arguments{
\item{selectionThreshold}{What sequence similarity percentage to the
original sequence we wish to maintain (distant sequences are dropped over the
course of the simulation)}

\item{speciesSimilarity}{If this proportion of the overall sequence similarity
matrix falls below a certain percentage (species_coehrence), then we split the
current species into two species, and allow for recombination only within
species.}

\item{speciesCoherence}{Refer to `speciesSimilarity`}
}
\description{
Create SpeciationParams object
}
