% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{BurstParams}
\alias{BurstParams}
\title{Create BurstParams object}
\usage{
BurstParams(burstProbability = 0.1, burstMean = 1, maxTotalCopies = 50)
}
\arguments{
\item{burstProbability}{The probability that an active retrotransposon will increase in copy number during a time jump of one timestep}

\item{burstMean}{The Poisson mean for the distribution that specifies how many new sequences an active sequence will create during bursting}

\item{maxTotalCopies}{The largest population size of sequences to keep track of (if this is exceeded, sequences are randomly discarded to simulate death)}
}
\value{
A bundling of the parameters given to it as a BurstParams object
}
\description{
Create BurstParams object
}
\examples{
burstParams <- BurstParams(burstMean = 2)
}
