% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{BurstParams}
\alias{BurstParams}
\title{Create BurstParams object}
\usage{
BurstParams(
  burstProbability = 0.1,
  burstMean = 1,
  recombMean = 1.5,
  recombSimilarity = 0.85
)
}
\arguments{
\item{burstProbability}{Probability that an active sequence bursts in one
timestep}

\item{burstMean}{The Poisson mean for the distribution that specifies how
many new sequences an active sequence will create during bursting}

\item{recombMean}{The expected number of template switches during
recombination between two sequences (chosen from a Poisson distribution with
this as its mean).}

\item{recombSimilarity}{TODO}
}
\description{
Create BurstParams object
}
