% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{OutputParams}
\alias{OutputParams}
\title{Create OutputParams object}
\usage{
OutputParams(
  outputFilename = "simulationOutput.out",
  outputNumRawSequence = 2,
  outputNumInitialDistance = 10,
  outputNumPairwiseDistance = 10,
  outputNumSpeciesTags = 10
)
}
\arguments{
\item{outputFilename}{Where should the results of the simulation be saved? (This
can be parsed by input_file)}

\item{outputNumRawSequence}{How many times across the simulation will we output the
raw sequences themselves}

\item{outputNumInitialDistance}{How many times across the simulation will we output the
distance of each sequence to the initial sequence}

\item{outputNumPairwiseDistance}{How many times across the simulation will we output the
pairwise distance between all pairs of sequences}

\item{outputNumSpeciesTags}{How many times across the simulation will we output
the tags (unique identifiers) of the sequences and their species}
}
\description{
Create OutputParams object
}
