% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file.R
\name{parseSimulationOutput}
\alias{parseSimulationOutput}
\title{Input data from CPP output into a set of R data frames}
\usage{
parseSimulationOutput(filename)
}
\arguments{
\item{filename}{The filename of the output generated by the simulation}
}
\description{
Input data from CPP output into a set of R data frames
}
