% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plotEvolution}
\alias{plotEvolution}
\title{Plotting function wrapper}
\usage{
plotEvolution(data, type, ...)
}
\arguments{
\item{data}{Simulation output data obtained from [retrocombinator::parseEvolutionOutput()()]}

\item{type}{What should be plotted?
\describe{
\item{"initial"}{Plot similarity to initial sequence over time.}
\item{"pairwise"}{Plot the similarity between sequences over time.}
\item{"families"}{Plot the evolution of families over time.}
}}

\item{...}{Any additional arguments to pass to specialised plotting functions
(currently unused)}
}
\value{
A corresponding evolutionary history plot, based on the simulation of
        the evolution of retrotransposons
}
\description{
Plotting function wrapper
}
\examples{
\dontrun{
plotEvolution(myData, "initial")
}
}
