% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plotEvolution}
\alias{plotEvolution}
\title{Plotting function wrapper}
\usage{
plotEvolution(data, type, ...)
}
\arguments{
\item{data}{Simulation output data obtained from parseEvolutionOutput}

\item{type}{"initial"  - Plot similarity to initial sequence over time.
"pairwise" - Plot the similarity between sequences over time.
"phylogeny" - Plot the evolution of families over time.}

\item{...}{Any additional arguments to pass to specialised plotting functions
(currently unused)}
}
\description{
Plotting function wrapper
}
