% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SeedParams}
\alias{SeedParams}
\title{Create SeedParams object}
\usage{
SeedParams(toRandomise = FALSE, toSeed = TRUE, seedForRandom = 0)
}
\arguments{
\item{toRandomise}{Should this simulation be run with a random seed to begin
with? (Based on system time)}

\item{toSeed}{Should this sequence be run with a specific initial seed? If
so, then the seed is specified by the parameter seedForRandom}

\item{seedForRandom}{Refer to documentation of toSeed}
}
\description{
Create SeedParams object
}
