% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SequenceParams}
\alias{SequenceParams}
\title{Create SequenceParams object}
\usage{
SequenceParams(num_initial_copies = 100, seq_length = 500, seq_filename = NULL)
}
\arguments{
\item{num_initial_copies}{The initial number of sequences to consider}

\item{seq_length}{The number of nucleotides in each of the sequences in the
simulation}

\item{seq_filename}{A path to a sequence to be used as initial sequence}
}
\description{
Create SequenceParams object
}
