% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEvolution.R
\name{simulateEvolution}
\alias{simulateEvolution}
\title{Run an entire simulation with flags}
\usage{
simulateEvolution(
  sequenceParams = SequenceParams(),
  simulationParams = SimulationParams(),
  mutationParams = MutationParams(),
  flagParams = FlagParams(),
  burstParams = BurstParams(),
  speciationParams = SpeciationParams(),
  outputParams = OutputParams(),
  seedParams = SeedParams()
)
}
\description{
Run an entire simulation with flags
}
