% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEvolution.R
\name{simulateEvolution}
\alias{simulateEvolution}
\title{Run an entire simulation with flags}
\usage{
simulateEvolution(
  sequenceParams = SequenceParams(),
  activityParams = ActivityParams(),
  mutationParams = MutationParams(),
  burstParams = BurstParams(),
  recombParams = RecombParams(),
  selectionParams = SelectionParams(),
  familyParams = FamilyParams(),
  simulationParams = SimulationParams(),
  outputParams = OutputParams(),
  seedParams = SeedParams()
)
}
\arguments{
\item{sequenceParams}{A [retrocombinator::SequenceParams()] object for the simulation}

\item{activityParams}{An [retrocombinator::ActivityParams()] object for the simulation}

\item{mutationParams}{A [retrocombinator::MutationParams()] object for the simulation}

\item{burstParams}{A [retrocombinator::BurstParams()] object for the simulation}

\item{recombParams}{A [retrocombinator::RecombParams()] object for the simulation}

\item{selectionParams}{A [retrocombinator::SelectionParams()] object for the simulation}

\item{familyParams}{A [retrocombinator::FamilyParams()] object for the simulation}

\item{simulationParams}{A [retrocombinator::SimulationParams()] object for the simulation}

\item{outputParams}{A [retrocombinator::OutputParams()] object for the simulation}

\item{seedParams}{A [retrocombinator::SeedParams()] object for the simulation}
}
\value{
A single character, the filename of the output of the simulation that
can then be loaded into R using [retrocombinator::parseSimulationOutput()]
}
\description{
Run an entire simulation with flags
}
\examples{
\dontrun{
simulateEvolution(sequenceParams = mySequenceParams,
                  familyParams = myFamilyParams)
}
}
