% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{SimulationParams}
\alias{SimulationParams}
\title{Create SimulationParams object}
\usage{
SimulationParams(numSteps = 20, timePerStep = 1)
}
\arguments{
\item{numSteps}{How many steps we have in our simulation}

\item{timePerStep}{How much time passes in one jump (unit: millions of years)}
}
\description{
Create SimulationParams object
}
